[ 1816524 ] Eliminate use of long integers in kernel code. [ 1816526 ] Cleanup char usage in the interpreter. [ 1816527 ] Eliminate OS/2-Windows type usage in portable code. [ 1823508 ] Update rexx.h to be more portable [ 1823508 ] Update rexx.h to be more portable
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@1499 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ kernel/classes/DirectoryClass.hpp
@@ -68,7 +68,7 @@ class RexxDirectory : public RexxHashTableCollection {
   RexxObject   *entryRexx(RexxString *);
   RexxObject   *hasEntry(RexxString *);
   RexxObject   *hasIndex(RexxString *);
-  long          items();
+  size_t        items();
   RexxObject   *itemsRexx();
   RexxObject   *remove(RexxString *);
   RexxObject   *setEntry(RexxString *, RexxObject *);
